Refactor IFolderContents interface into its own file.
Now all interfaces and classes are in separate files which makes the
source code more consistently organized.
